Works with PETSc now

+++ parallel/parallel_src/interface/parhip_interface.h
@@ -7,12 +7,13 @@
 #ifndef PARHIP_INTERFACE
 #define PARHIP_INTERFACE
 #include <mpi.h>
+#include <stdbool.h>
 #ifdef __cplusplus
 
 extern "C"
 {
 #endif
-typedef unsigned long long idxtype;
+typedef int idxtype;
 
 const int ULTRAFASTMESH   = 0;
 const int FASTMESH        = 1;
@@ -22,7 +23,7 @@ const int FASTSOCIAL      = 4;
 const int ECOSOCIAL       = 5;
 
 void ParHIPPartitionKWay(idxtype *vtxdist, idxtype *xadj, idxtype *adjncy, idxtype *vwgt, idxtype *adjwgt,
-                         int *nparts, double* imbalance, bool suppress_output, int seed, int mode, int *edgecut, idxtype *part, 
+                         int *nparts, double* imbalance, bool suppress_output, int seed, int mode, int *edgecut, idxtype *part,
                          MPI_Comm *comm);
 #ifdef __cplusplus
 }
